Add plot base and move band / dos plot there
