+++ CRYSTALpytools/base/__init__.py
@@ -3,3 +3,4 @@ from CRYSTALpytools.base import inputbase
 from CRYSTALpytools.base import crysd12
 from CRYSTALpytools.base import crysout
 from CRYSTALpytools.base import propout
+from CRYSTALpytools.base import plotbase
